Move earcut test data
