+++ modules/polygon/test/data/earcut/README.md
@@ -1,0 +1,3 @@
+# Earcut test data
+
+Test data from original [Earcut library tests](https://github.com/mapbox/earcut/tree/master/test), licensed under [ISC license](https://github.com/mapbox/earcut/blob/master/LICENSE).
