+++ modules/polygon/test/data/earcut/expected.js
