+++ modules/polygon/test/data/earcut/fixtures/bad-diagonals.json
