+++ modules/polygon/test/data/earcut/fixtures/bad-hole.json
