+++ modules/polygon/test/data/earcut/fixtures/boxy.json
