+++ modules/polygon/test/data/earcut/fixtures/building.json
