+++ modules/polygon/test/data/earcut/fixtures/collinear-diagonal.json
