+++ modules/polygon/test/data/earcut/fixtures/degenerate.json
