+++ modules/polygon/test/data/earcut/fixtures/dude.json
