+++ modules/polygon/test/data/earcut/fixtures/eberly-3.json
