+++ modules/polygon/test/data/earcut/fixtures/eberly-6.json
