+++ modules/polygon/test/data/earcut/fixtures/empty-square.json
