+++ modules/polygon/test/data/earcut/fixtures/hole-touching-outer.json
