+++ modules/polygon/test/data/earcut/fixtures/hourglass.json
