+++ modules/polygon/test/data/earcut/fixtures/issue107.json
