+++ modules/polygon/test/data/earcut/fixtures/issue111.json
