+++ modules/polygon/test/data/earcut/fixtures/issue119.json
