+++ modules/polygon/test/data/earcut/fixtures/issue131.json
