+++ modules/polygon/test/data/earcut/fixtures/issue16.json
