+++ modules/polygon/test/data/earcut/fixtures/issue17.json
