+++ modules/polygon/test/data/earcut/fixtures/issue29.json
