+++ modules/polygon/test/data/earcut/fixtures/issue34.json
