+++ modules/polygon/test/data/earcut/fixtures/issue35.json
