+++ modules/polygon/test/data/earcut/fixtures/issue45.json
