+++ modules/polygon/test/data/earcut/fixtures/issue52.json
