+++ modules/polygon/test/data/earcut/fixtures/issue83.json
