+++ modules/polygon/test/data/earcut/fixtures/outside-ring.json
