+++ modules/polygon/test/data/earcut/fixtures/rain.json
