+++ modules/polygon/test/data/earcut/fixtures/self-touching.json
