+++ modules/polygon/test/data/earcut/fixtures/shared-points.json
