+++ modules/polygon/test/data/earcut/fixtures/simplified-us-border.json
