+++ modules/polygon/test/data/earcut/fixtures/steiner.json
