+++ modules/polygon/test/data/earcut/fixtures/touching-holes.json
