+++ modules/polygon/test/data/earcut/fixtures/touching2.json
