+++ modules/polygon/test/data/earcut/fixtures/touching3.json
