+++ modules/polygon/test/data/earcut/fixtures/touching4.json
