+++ modules/polygon/test/data/earcut/fixtures/water-huge.json
