+++ modules/polygon/test/data/earcut/fixtures/water-huge2.json
