+++ modules/polygon/test/data/earcut/fixtures/water.json
