+++ modules/polygon/test/data/earcut/fixtures/water2.json
