+++ modules/polygon/test/data/earcut/fixtures/water3.json
