+++ modules/polygon/test/data/earcut/fixtures/water3b.json
