+++ modules/polygon/test/data/earcut/fixtures/water4.json
